fix 3 issues
